python: Move dazl.model.reading to dazl.protocols.events.
