+++ python/dazl/client/bots.py
@@ -29,9 +29,9 @@ from typing import (
 from uuid import uuid4
 
 from .. import LOG
-from ..model.reading import BaseEvent
 from ..model.writing import Command, CommandBuilder
 from ..prim import Party
+from ..protocols.events import BaseEvent
 from ..util.asyncio_util import LongRunningAwaitable, Signal, completed, failed, propagate
 from .events import EventKey
