+++ python/tests/profiling/test_memory.py
@@ -17,7 +17,7 @@ import sys
 from pympler import muppy, summary
 
 from dazl import Network, create
-from dazl.model.reading import ContractCreateEvent, ReadyEvent
+from dazl.protocols.events import ContractCreateEvent, ReadyEvent
 
 
 def main(url: str):
