+++ python/tests/unit/test_api_consistency.py
@@ -1,8 +1,8 @@
 # Copyright (c) 2017-2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
+from dazl import Party
 from dazl.client.api import AIOPartyClient, SimplePartyClient
-from dazl.model.core import Party
 
 
 def test_api_consistency():
